// types.ts
export interface DecalogueItemData {
  id: number;
  title: string;
  description: string;
  iconClass: string; // Font Awesome class e.g. "fa-solid fa-sitemap"
  textBgClass: string; // Tailwind class for text area background e.g. "bg-sky-100"
  textFgClass: string; // Tailwind class for text area foreground e.g. "text-sky-800"
  iconBgClass: string; // Tailwind class for icon area background e.g. "bg-sky-500"
  iconFgClass: string; // Tailwind class for icon area foreground e.g. "text-white"
}